- 存在しないplace_idの対応 - prettier修正
