+++ src/Place/Endpoint/PlaceHttpPost.ts
@@ -3,27 +3,26 @@ import { ENV } from '@/ENV';
 export const PlaceHttpPost = async <RequestBodyType, ResponseType>(
 	target: string,
 	body: RequestBodyType,
-): Promise<ResponseType | unknown> => {
-	try {
-		const res = await fetch(ENV.BACKEND_PLACE_ENDPOINT + target, {
+): Promise<ResponseType> => new Promise<ResponseType>(async (resolve, reject) => {
+		const res = (await fetch(ENV.BACKEND_PLACE_ENDPOINT + target, {
 			method: 'POST',
 			headers: {
 				Accept: 'application/json',
 				'Content-Type': 'application/json',
 			},
 			body: JSON.stringify(body),
-		});
-		const resJSON = (await res.json()) as ResponseType;
+		}).catch((e) => {
+			reject(e);
+		})) as Response;
+		if (!res.ok) {
+			reject();
+		}
+		const resJSON = (await res.json().catch((e) => {
+			reject(e);
+		})) as ResponseType;
 		if (ENV.HTTP_POST_LOG) {
 			// eslint-disable-next-line no-console
 			console.log(resJSON);
 		}
-		return resJSON;
-	} catch (e) {
-		if (ENV.HTTP_POST_LOG) {
-			// eslint-disable-next-line no-console
-			console.error(e);
-		}
-		return e;
-	}
-};
+		resolve(resJSON);
+	});
