+++ src/Place/Models/PDB01MPlace/Contexts/PCT012MPlaceOne/ModelComponents/PMC01201GoogleMapPlaceOne/GoogleMapPlaceOne.web.tsx
@@ -9,18 +9,18 @@ export const PMC01201GoogleMapPlaceOne = () => {
 	// TODO: https://github.com/Ayato-kosaka/spelieve/issues/274 マップ範囲調整
 	const { place } = useContext(PCT012MPlaceOne);
 
-	if (place) {
-		const center = {
-			lat: place.geometry.latitude,
-			lng: place.geometry.longitude,
-		};
-
-		const coordinate = { lat: place.geometry.latitude, lng: place.geometry.longitude };
-		return (
-			<GoogleMap mapContainerStyle={styles.containerStyle} center={center} zoom={10}>
-				<MarkerWeb position={coordinate} title={place.name} key={place.place_id} />;
-			</GoogleMap>
-		);
+	if (!place) {
+		return <></>;
 	}
-	return <></>;
+	const center = {
+		lat: place.geometry.latitude,
+		lng: place.geometry.longitude,
+	};
+
+	const coordinate = { lat: place.geometry.latitude, lng: place.geometry.longitude };
+	return (
+		<GoogleMap mapContainerStyle={styles.containerStyle} center={center} zoom={10}>
+			<MarkerWeb position={coordinate} title={place.name} key={place.place_id} />;
+		</GoogleMap>
+	);
 };
