+++ src/Place/Pages/PPA002Place/Place.tsx
@@ -20,8 +20,7 @@ import { PCT012MPlaceOne } from '@/Place/Models/PDB01MPlace/Contexts/PCT012MPlac
 export const PPA002Place = ({ route, navigation }: NativeStackScreenProps<BottomTabParamList, 'PPA002Place'>) => {
 	const { place, setPlaceID, isLoading } = useContext(PCT012MPlaceOne);
 	const { place_id, language } = route.params;
-	const { onCreateItineraryClicked, displayOpeningHours } = PPA002PlaceController(route.params);
-	let openingInfo;
+	const { onCreateItineraryClicked } = PPA002PlaceController(route.params);
 
 	useEffect(() => {
 		setPlaceID(place_id);
@@ -31,103 +30,99 @@ export const PPA002Place = ({ route, navigation }: NativeStackScreenProps<Bottom
 	if (isLoading) {
 		return <ActivityIndicator animating />;
 	}
-	if (!isLoading && !place) {
-		return <Text>{i18n.t('Place Not Found')}</Text>;
+	if (!place) {
+		return <Text style={styles.infoText}>{i18n.t('Place Not Found')}</Text>;
 	}
-	if (!isLoading && place) {
-		openingInfo = displayOpeningHours(place.openingHours);
 
-		return (
-			<ScrollView>
-				<View style={styles.container}>
+	return (
+		<ScrollView>
+			<View style={styles.container}>
+				<View>
+					<PMC01201GoogleMapPlaceOne />
+					<PMC01203PlaceImage />
+					<Text style={styles.infoText}>{place.name}</Text>
+					<BookIcon name="open-book" size={50} onPress={() => onCreateItineraryClicked(place.name)} />
+					<FetherIcon name="map-pin" size={20}>
+						<Text style={styles.infoText}>{place.formatted_address}</Text>
+					</FetherIcon>
+					<WebIcon name="web" size={20}>
+						<Text
+							style={styles.urlLink}
+							onPress={() => {
+								// eslint-disable-next-line
+								Linking.openURL(`${place.website || ''}`);
+							}}>
+							{place.website ? place.website : i18n.t('No Web Infomation')}
+						</Text>
+					</WebIcon>
+					<FetherIcon name="phone-call" size={20}>
+						<Text style={styles.infoText}>
+							{place.formatted_phone_number ? place.formatted_phone_number : i18n.t('No Tel Information')}
+						</Text>
+					</FetherIcon>
 					<View>
-						<PMC01201GoogleMapPlaceOne />
-						<PMC01203PlaceImage />
-						<Text style={styles.infoText}>{place.name}</Text>
-						<BookIcon name="open-book" size={50} onPress={() => onCreateItineraryClicked(place.name)} />
-						<FetherIcon name="map-pin" size={20}>
-							<Text style={styles.infoText}>{place.formatted_address}</Text>
-						</FetherIcon>
-						<WebIcon name="web" size={20}>
-							<Text
-								style={styles.urlLink}
-								onPress={() => {
-									// eslint-disable-next-line
-									Linking.openURL(`${place.website || ''}`);
-								}}>
-								{place.website ? place.website : i18n.t('No Web Infomation')}
-							</Text>
-						</WebIcon>
-						<FetherIcon name="phone-call" size={20}>
-							<Text style={styles.infoText}>
-								{place.formatted_phone_number ? place.formatted_phone_number : i18n.t('No Tel Information')}
-							</Text>
-						</FetherIcon>
-						<View>
-							<TimeIcon name="time-outline" size={20}>
-								<Text style={styles.infoText}>{i18n.t('Opening Hours')}</Text>
-							</TimeIcon>
-							{Array.isArray(openingInfo) ? (
-								<FlatList
-									data={openingInfo}
-									renderItem={(itemData) => {
-										const [day, time] = itemData.item;
-										return (
-											<View>
-												<Text style={styles.infoText}>
-													{day} {time}
-												</Text>
-											</View>
-										);
-									}}
-									numColumns={1}
-								/>
-							) : (
-								<Text style={styles.infoText}>{openingInfo}</Text>
-							)}
-							<Text
-								style={styles.urlLink}
-								onPress={() => {
-									// eslint-disable-next-line
-									Linking.openURL(`${place.mapUrl}`);
-								}}>
-								{i18n.t('show more')}
-							</Text>
-						</View>
-					</View>
-					<View>
-						<Text style={styles.infoText}>{i18n.t('Customer Reviews')}</Text>
-						<Rating type="star" readonly jumpValue={0.1} startingValue={place.rating} />
+						<TimeIcon name="time-outline" size={20}>
+							<Text style={styles.infoText}>{i18n.t('Opening Hours')}</Text>
+						</TimeIcon>
+						{Array.isArray(place.openingHours) ? (
+							<FlatList
+								data={place.openingHours}
+								renderItem={(itemData) => {
+									const [day, time] = itemData.item;
+									return (
+										<View>
+											<Text style={styles.infoText}>
+												{day} {time}
+											</Text>
+										</View>
+									);
+								}}
+								numColumns={1}
+							/>
+						) : (
+							<Text style={styles.infoText}>{place.openingHours}</Text>
+						)}
 						<Text
 							style={styles.urlLink}
 							onPress={() => {
 								// eslint-disable-next-line
 								Linking.openURL(`${place.mapUrl}`);
-							}}
-						/>
-					</View>
-					<View>
-						<FlatList
-							data={place.photoUrls}
-							renderItem={(itemData) => (
-								<View>
-									<Image source={{ uri: itemData.item }} style={styles.image} />
-								</View>
-							)}
-							numColumns={3}
-							// keyExtractor={(place) => place.place_id}
-						/>
-						<Text
-							style={styles.urlLink}
-							onPress={() => {
-								// eslint-disable-next-line
-								Linking.openURL(`${place.mapUrl}`);
-							}}
-						/>
+							}}>
+							{i18n.t('show more')}
+						</Text>
 					</View>
 				</View>
-			</ScrollView>
-		);
-	}
-	return <></>;
+				<View>
+					<Text style={styles.infoText}>{i18n.t('Customer Reviews')}</Text>
+					<Rating type="star" readonly jumpValue={0.1} startingValue={place.rating} />
+					<Text
+						style={styles.urlLink}
+						onPress={() => {
+							// eslint-disable-next-line
+							Linking.openURL(`${place.mapUrl}`);
+						}}
+					/>
+				</View>
+				<View>
+					<FlatList
+						data={place.photoUrls}
+						renderItem={(itemData) => (
+							<View>
+								<Image source={{ uri: itemData.item }} style={styles.image} />
+							</View>
+						)}
+						numColumns={3}
+						// keyExtractor={(place) => place.place_id}
+					/>
+					<Text
+						style={styles.urlLink}
+						onPress={() => {
+							// eslint-disable-next-line
+							Linking.openURL(`${place.mapUrl}`);
+						}}
+					/>
+				</View>
+			</View>
+		</ScrollView>
+	);
 };
