+++ src/Place/Pages/PPA002Place/PlaceController.ts
@@ -1,14 +1,9 @@
 import { useNavigation } from '@react-navigation/native';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack/lib/typescript/src/types';
 
-import {
-	MPlaceOpeningHoursInterface,
-	PlaceControllerInterface,
-	PlacePropsInterface,
-} from 'spelieve-common/lib/Interfaces';
+import { PlaceControllerInterface, PlacePropsInterface } from 'spelieve-common/lib/Interfaces';
 
 import { BottomTabParamList } from '@/App';
-import i18n from '@/Common/Hooks/i18n-js';
 
 export const PPA002PlaceController = ({ place_id, language }: PlacePropsInterface): PlaceControllerInterface => {
 	const navigation = useNavigation<NativeStackNavigationProp<BottomTabParamList>>();
@@ -20,43 +15,8 @@ export const PPA002PlaceController = ({ place_id, language }: PlacePropsInterfac
 		});
 	};
 
-	const displayOpeningHours = (
-		openingHours: MPlaceOpeningHoursInterface[] | undefined,
-	): string | Array<[string, string]> => {
-		if (!openingHours) {
-			return i18n.t('No Opening Hours Infomation');
-		}
-
-		if (openingHours.length === 1 && !openingHours[0].close) {
-			return i18n.t('Open 24hours');
-		}
-
-		const days: { [key: number]: string } = {
-			0: 'Sunday',
-			1: 'Monday',
-			2: 'Tuesday',
-			3: 'Wednesday',
-			4: 'Thursday',
-			5: 'Friday',
-			6: 'Saturday',
-		};
-		const changeTimeView = (time: string): string => {
-			// 1300->13:00
-			const hours = time.slice(0, 2);
-			const minutes = time.slice(2, 4);
-			return `${hours}:${minutes}`;
-		};
-		return openingHours.map((openingHour) => {
-			const { open } = openingHour;
-			const { close } = openingHour;
-			const day = days[open.day];
-			const time = `${changeTimeView(open.time)}~${changeTimeView(close.time)}`;
-			return [i18n.t(day), time];
-		});
-	};
-
 	// TODO: https://github.com/Ayato-kosaka/spelieve/issues/320 画像押したらモーダルで開くようにする
 	const onImageClicked = () => {};
 
-	return { onCreateItineraryClicked, displayOpeningHours, onImageClicked };
+	return { onCreateItineraryClicked, onImageClicked };
 };
